Add files via upload
These days, we see that a lot of individuals prefer to shop online rather than visiting physical places to make their purchases. We have seen that a single internet product can have thousands or many thousands of customer evaluations, making it challenging to evaluate the product. Through this project, we are creating a graph-based model out of the customer feedback. This facilitates the purchasing process for customers and makes it simple for retailers. Sentiment analysis is what we utilize in the project to examine the reviews. It determines if the review is positive, negative or neutral before plotting the data and uses several machine learning algorithms for increased accuracy.

DEVELOPMENT PROCESS:

Step-I:
	Open chrome and add extension amazon review export to your chrome. Then login to it using your credentials.

Step-II:
	Open amazon and open reviews page.

Step-III:
	Now click on extension and click on export reviews.

Step-IV:
	Now in the second dataset place use this exported reviews dataset in the code.